test: 🗜 replace MockContext by a fake context and regroup test helpers
