+++ packages/semver/src/builders/version/builder.e2e.spec.ts
@@ -1,12 +1,11 @@
 import { BuilderOutput } from '@angular-devkit/architect';
 import { execSync } from 'child_process';
-import { mkdirSync, writeFileSync } from 'fs';
-import { dirname, resolve } from 'path';
-import * as rimraf from 'rimraf';
-
-import * as tmp from 'tmp';
-import { promisify } from 'util';
 import { runBuilder } from './builder';
+import {
+  createFakeContext,
+  setupTestingWorkspace,
+  TestingWorkspace,
+} from './testing';
 import { readPackageJson } from './utils/project';
 
 describe('@jscutlery/semver:version e2e', () => {
@@ -69,6 +68,7 @@ describe('@jscutlery/semver:version e2e', () => {
         createFakeContext({
           project: 'workspace',
           projectRoot: testingWorkspace.root,
+          workspaceRoot: testingWorkspace.root,
         })
       ).toPromise();
     });
@@ -83,61 +83,4 @@ describe('@jscutlery/semver:version e2e', () => {
       ).toEqual('0.1.0');
     });
   });
-
-  interface TestingWorkspace {
-    tearDown(): Promise<void>;
-    root: string;
-  }
-
-  function setupTestingWorkspace(files: Map<string, string>): TestingWorkspace {
-    /* Create a temporary directory. */
-    const tmpDir = tmp.dirSync();
-
-    for (const [fileRelativePath, content] of files.entries()) {
-      const filePath = resolve(tmpDir.name, fileRelativePath);
-      const directory = dirname(filePath);
-      /* Create path. */
-      mkdirSync(directory, { recursive: true });
-      /* Create file. */
-      writeFileSync(filePath, content, 'utf-8');
-    }
-
-    const originalCwd = process.cwd();
-    process.chdir(tmpDir.name);
-
-    /* Retrieving path from `process.cwd()`
-     * because for some strange reasons it returns a different value.
-     * Cf. https://github.com/nodejs/node/issues/7545 */
-    const workspaceRoot = process.cwd();
-
-    return {
-      /**
-       * Destroy and restore cwd.
-       */
-      async tearDown() {
-        await promisify(rimraf)(workspaceRoot);
-        process.chdir(originalCwd);
-      },
-      root: workspaceRoot,
-    };
-  }
-
-  function createFakeContext({
-    project,
-    projectRoot,
-  }: {
-    project: string;
-    projectRoot: string;
-  }) {
-    return {
-      getProjectMetadata: jest.fn().mockReturnValue({ root: projectRoot }),
-      logger: { error: jest.fn() },
-      reportStatus: jest.fn(),
-      target: {
-        project,
-      },
-      workspaceRoot: testingWorkspace.root,
-      /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
-    } as any;
-  }
 });
